Fixed some (hopefully all) dgmg bugs. 1. choose destination picked last src node despite being given src id. 2. node_hv representation was not actually being applied to the nodes (thought it edited in place. not true for for-loops I guess). Also used the same loss function as choose dest scoring method, log_softmax, vs. nn.CrossEntropyLoss, just in case.... And other edits.

+++ housingpipeline/housingpipeline/dgmg/utils.py
@@ -284,7 +284,7 @@ def one_hot_embedding(labels, num_classes=11):
     """
     y = torch.eye(num_classes)
     # print(" label is",labels)
-    return y[labels.int()]
+    return y[labels.to(dtype=torch.long)]
 
 
 def dgl_to_graphlist(g):
